+++ sentry/src/main/java/io/sentry/Span.java
@@ -106,19 +106,39 @@ public final class Span implements ISpan {
   }
 
   @Override
+  public @Nullable TraceState traceState() {
+    return transaction.traceState();
+  }
+
+  @Override
+  public @Nullable TraceStateHeader toTraceStateHeader() {
+    return transaction.toTraceStateHeader();
+  }
+
+  @Override
   public void finish() {
     this.finish(this.context.getStatus());
   }
 
   @Override
   public void finish(@Nullable SpanStatus status) {
+    finish(status, DateUtils.getCurrentDateTime());
+  }
+
+  /**
+   * Used to finish unfinished spans by {@link SentryTracer}.
+   *
+   * @param status - status to finish span with
+   * @param timestamp - the root span timestamp.
+   */
+  void finish(@Nullable SpanStatus status, Date timestamp) {
     // the span can be finished only once
     if (!finished.compareAndSet(false, true)) {
       return;
     }
 
     this.context.setStatus(status);
-    timestamp = DateUtils.getCurrentDateTime();
+    this.timestamp = timestamp;
     if (throwable != null) {
       hub.setSpanContext(throwable, this, this.transaction.getName());
     }